Use signed arithmetic in mutt_gecos_name to avoid an overflow. Closes #2885.

+++ muttlib.c
@@ -540,7 +540,7 @@ char *mutt_gecos_name (char *dest, size_t destlen, struct passwd *pw)
     if (dest[idx] == '&')
     {
       memmove (&dest[idx + pwnl], &dest[idx + 1],
-	       MAX(destlen - idx - pwnl - 1, 0));
+	       MAX((ssize_t)(destlen - idx - pwnl - 1), 0));
       memcpy (&dest[idx], pw->pw_name, MIN(destlen - idx - 1, pwnl));
       dest[idx] = toupper ((unsigned char) dest[idx]);
     }
